requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-oauth2==4.1
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.2.0
graphene-federation==3.1.4
graphql-core==3.2.3
graphql-relay==3.2.0
Markdown==3.5.2
mysqlclient==2.2.1
promise==2.3
pytz==2024.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
